Move selectors under state/automated-transfer.
Also reorganises a few of the existing ones.
