+++ client/state/automated-transfer/selectors/get-automated-transfer.ts
@@ -1,0 +1,12 @@
+/**
+ * External dependencies
+ */
+import { get } from 'lodash';
+
+/**
+ * Internal dependencies
+ */
+import 'state/automated-transfer/init';
+
+export const getAutomatedTransfer = ( state, siteId: number | null ) =>
+	get( state, [ 'automatedTransfer', siteId ], {} );
