+++ client/state/automated-transfer/selectors/index.ts
@@ -6,27 +6,18 @@ import { flowRight as compose, get } from 'lodash';
 /**
  * Internal dependencies
  */
+import { getAutomatedTransfer } from './get-automated-transfer';
+
 import 'state/automated-transfer/init';
 
-export const getAutomatedTransfer = ( state, siteId: number | null ) =>
-	get( state, [ 'automatedTransfer', siteId ], {} );
-
 /**
- * Helper to get status state from local transfer state sub-tree
- *
- * @param {object} state automated transfer state sub-tree for a site
- * @returns {string} status of transfer
+ * Re-exports
  */
-export const getStatusData = state => get( state, 'status', null );
-
-/**
- * Returns status info for transfer
- *
- * @param {object} state global app state
- * @param {number} siteId requested site for transfer info
- * @returns {string|null} status if available else `null`
- */
-export const getAutomatedTransferStatus = compose( getStatusData, getAutomatedTransfer );
+export { getAutomatedTransfer } from './get-automated-transfer';
+export { getAutomatedTransferStatus } from './get-automated-transfer-status';
+export { isAutomatedTransferActive } from './is-automated-transfer-active';
+export { isAutomatedTransferFailed } from './is-automated-transfer-failed';
+export { default as isFetchingAutomatedTransferStatus } from './is-fetching-automated-transfer-status';
 
 export interface EligibilityWarning {
 	description: string;
