+++ client/state/automated-transfer/selectors/is-fetching-automated-transfer-status.js
@@ -6,7 +6,7 @@ import { get, isEmpty } from 'lodash';
 /**
  * Internal Dependencies
  */
-import { getAutomatedTransfer } from 'state/automated-transfer/selectors';
+import { getAutomatedTransfer } from './get-automated-transfer';
 
 import 'state/automated-transfer/init';
