+++ client/state/selectors/test/is-automated-transfer-active.js
@@ -6,7 +6,7 @@ import { expect } from 'chai';
 /**
  * Internal dependencies
  */
-import { isActive } from '../is-automated-transfer-active';
+import { isActive } from 'state/automated-transfer/selectors/is-automated-transfer-active';
 import { transferStates } from 'state/automated-transfer/constants';
 
 describe( 'Automated Transfer', () => {
