+++ client/state/selectors/test/is-automated-transfer-failed.js
@@ -6,7 +6,7 @@ import { expect } from 'chai';
 /**
  * Internal dependencies
  */
-import { isFailed } from '../is-automated-transfer-failed';
+import { isFailed } from 'state/automated-transfer/selectors/is-automated-transfer-failed';
 import { transferStates } from 'state/automated-transfer/constants';
 
 describe( 'Automated Transfer', () => {
